pyproject.toml
[project]
name = "enem-backend"
version = "0.1.0"
description = "Backend profissional do sistema de correção ENEM com Django 5.1 + DRF + JWT"
authors = [
    {name = "Soeiro", email = "mandre.soeiro@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django==5.1",
    "djangorestframework",
    "djangorestframework-simplejwt",
    "dj-rest-auth",
    "django-allauth",
    "python-decouple",
    "psycopg2-binary",
    "weasyprint",
    "drf-spectacular",
    "whitenoise",
    "requests (>=2.32.5,<3.0.0)",
    "django-jazzmin (>=3.0.1,<4.0.0)"
]

[tool.poetry]
packages = [
    { include = "core" }
]

[tool.poetry.group.dev.dependencies]
black = "*"
ruff = "*"
isort = "*"
pytest = "*"
pytest-django = "*"
mypy = "*"

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.isort]
profile = "black"

[tool.ruff]
line-length = 88
target-version = "py312"
select = ["E", "W", "F"]
ignore = ["E501"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
